+++ client/src/App.jsx
@@ -1,8 +1,12 @@
+import { Analytics } from '@vercel/analytics/react';
 import Home from './pages/Home';
 
 function App() {
   return (
-    <Home />
+    <>
+      <Home />
+      <Analytics />
+    </>
   );
 }